requirements.txt
pandas
sqlalchemy
pyodbc
azure-storage-blob
python-dotenv